refactor(company-verification): move CompanyVerification type and mock data to AdminService
