+++ src/features/admin/hooks/useCompanyVerification.ts
@@ -1,110 +1,22 @@
 'use client';
 
 import { useState, useEffect, useCallback } from 'react';
-
-export interface CompanyVerification {
-  id: number;
-  name: string;
-  location: string;
-  industry: string;
-  contact: string;
-  submitted: string;
-  status: 'pending' | 'rejected' | 'approved';
-}
-
-// Mock data matching the screenshot
-const mockRejectedCompanies: CompanyVerification[] = [
-  {
-    id: 1,
-    name: 'Tech Solutions Co.',
-    location: 'Bangkok, Thailand',
-    industry: 'Software Development',
-    contact: 'contact@techsolutions.com',
-    submitted: '2025-09-30',
-    status: 'rejected',
-  },
-  {
-    id: 2,
-    name: 'Digital Marketing Hub',
-    location: 'Chiang Mai, Thailand',
-    industry: 'Marketing',
-    contact: 'hr@digitalhub.co.th',
-    submitted: '2025-09-30',
-    status: 'rejected',
-  },
-  {
-    id: 3,
-    name: 'Tech Solutions Co.',
-    location: 'Bangkok, Thailand',
-    industry: 'Software Development',
-    contact: 'contact@techsolutions.com',
-    submitted: '2025-09-30',
-    status: 'rejected',
-  },
-  {
-    id: 4,
-    name: 'Digital Marketing Hub',
-    location: 'Chiang Mai, Thailand',
-    industry: 'Marketing',
-    contact: 'hr@digitalhub.co.th',
-    submitted: '2025-09-30',
-    status: 'rejected',
-  },
-  {
-    id: 5,
-    name: 'Tech Solutions Co.',
-    location: 'Bangkok, Thailand',
-    industry: 'Software Development',
-    contact: 'contact@techsolutions.com',
-    submitted: '2025-09-30',
-    status: 'rejected',
-  },
-  {
-    id: 6,
-    name: 'Digital Marketing Hub',
-    location: 'Chiang Mai, Thailand',
-    industry: 'Marketing',
-    contact: 'hr@digitalhub.co.th',
-    submitted: '2025-09-30',
-    status: 'rejected',
-  },
-  {
-    id: 7,
-    name: 'Tech Solutions Co.',
-    location: 'Bangkok, Thailand',
-    industry: 'Software Development',
-    contact: 'contact@techsolutions.com',
-    submitted: '2025-09-30',
-    status: 'rejected',
-  },
-];
+import { AdminService, type CompanyVerification } from '@/lib/services';
 
 export function useCompanyVerification() {
   const [companies, setCompanies] = useState<CompanyVerification[]>([]);
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
 
-  const fetchCompanies = useCallback(async (status?: 'pending' | 'rejected' | 'approved') => {
+  const fetchCompanies = useCallback(async () => {
     setIsLoading(true);
     setError(null);
 
     try {
-      // Simulate API call
-      await new Promise((resolve) => setTimeout(resolve, 700));
-
-      // In a real app, you would fetch from API:
-      // const response = await fetch(`/api/admin/companies?status=${status}`);
-      // const data = await response.json();
-      // setCompanies(data);
-
-      // For now, use mock data
-      const filtered = status
-        ? mockRejectedCompanies.filter((c) => c.status === status)
-        : mockRejectedCompanies;
-
-      setCompanies(filtered);
+      const data = await AdminService.getRejectedCompanies();
+      setCompanies(data);
     } catch (err) {
-      console.error('Error fetching companies:', err);
+      console.error('Error fetching rejected companies:', err);
       setError(err instanceof Error ? err.message : 'Failed to fetch companies');
     } finally {
       setIsLoading(false);
@@ -112,48 +24,16 @@ export function useCompanyVerification() {
   }, []);
 
   useEffect(() => {
-    // Fetch rejected companies by default
-    fetchCompanies('rejected');
+    fetchCompanies();
   }, [fetchCompanies]);
 
   const reconsiderCompany = async (companyId: number) => {
     try {
-      // In a real app:
-      // await fetch(`/api/admin/companies/${companyId}/reconsider`, { method: 'POST' });
-      console.log('Reconsidering company:', companyId);
-      await fetchCompanies('rejected'); // Refresh list
+      await AdminService.reconsiderCompany(companyId);
+      await fetchCompanies(); // Refresh list
       return true;
     } catch (err) {
       console.error('Error reconsidering company:', err);
-      throw err;
-    }
-  };
-
-  const approveCompany = async (companyId: number) => {
-    try {
-      // In a real app:
-      // await fetch(`/api/admin/companies/${companyId}/approve`, { method: 'POST' });
-      console.log('Approving company:', companyId);
-      await fetchCompanies('rejected'); // Refresh list
-      return true;
-    } catch (err) {
-      console.error('Error approving company:', err);
-      throw err;
-    }
-  };
-
-  const rejectCompany = async (companyId: number, reason?: string) => {
-    try {
-      // In a real app:
-      // await fetch(`/api/admin/companies/${companyId}/reject`, {
-      //   method: 'POST',
-      //   body: JSON.stringify({ reason })
-      // });
-      console.log('Rejecting company:', companyId, 'Reason:', reason);
-      await fetchCompanies('rejected'); // Refresh list
-      return true;
-    } catch (err) {
-      console.error('Error rejecting company:', err);
       throw err;
     }
   };
@@ -164,7 +44,7 @@ export function useCompanyVerification() {
     error,
     refetch: fetchCompanies,
     reconsiderCompany,
-    approveCompany,
-    rejectCompany,
   } as const;
 }
+
+export type { CompanyVerification };
