+++ src/lib/services/index.ts
@@ -10,4 +10,4 @@ export { JobService } from './job.service';
 export { CompanyService } from './company.service';
 export { AdminService } from './admin.service';
 
-export type { Report } from './admin.service';
+export type { Report, CompanyVerification } from './admin.service';
